First implementation of the driver abstraction for joysticks
Added the abstract class JoystickDriver. There is still only one driver
implementation for Linux joysticks (which is hardcoded to /dev/input/

The class Joystick was changed to only provide a description for a
Joystick (e.g. name) including its capabilities (number of buttons and
axis).

This version only prints the joystick event and quits after 10 seconds.
This was done to check whether stop works as intended. It does not send
MIDI events anymore.