Reinstating the functionality to send MIDI events
A new class JoystickToMIDIMapper was added which is intended to later
map Joystick events to MIDI events. It is given to the JoystickDriver
when JoystickDriver::start is called. Currently it only contains methods
like processJoystickEventButtonDown and processJoystickEventButtonUp
which will be replaced by something more generic and useful later.

Other changes:
- Decreased the sleeping time in the Linux joystick driver.
- Added an endless loop around the sleeping section in main to be able
to test whether MIDI events arrive.