Driver factories, null drivers and drivers selectable from GUI
Driver factories for joystick and MIDI drivers were added. Furthermore a
null driver that simply does nothing was added for both driver types.
Drivers are now selectable from a menu in the GUI.

The Linux kernel joystick driver does not throw an exception anymore
when the joystick is not plugged in, i.e. when /dev/input/js0 cannot be
opened. Instead the driver will stay uninitialized and report 0
joysticks.

Changed the interface of the joystick driver to return a collection of
joysticks.

Renamed MIDIDriverAlsa to MIDIDriverALSA.